Rename module: search_util
