+++ biomass/estimation/__init__.py
@@ -1,2 +1,2 @@
 from .ga import GeneticAlgorithmContinue, GeneticAlgorithmInit
-from .util import convert_scale, initialize_search_param
+from .search_util import convert_scale, initialize_search_param
